Calculate and print the actual cutoff percentage.
For optimized E*, this can differ between sampling repetitions,
so the cutoff percentage (`cut%`) exhibits non-zero standard
deviation.

For E* determined by a given threshold, this should just reproduce
the threshold value, so it serves as a double check.

Also, when reading a trajectory, check that it has enough samples
as requested.